+++ NINA.Profile/Interfaces/ITelescopeSettings.cs
@@ -1,7 +1,7 @@
 #region "copyright"
 
 /*
-    Copyright © 2016 - 2024 Stefan Berg <isbeorn86+NINA@googlemail.com> and the N.I.N.A. contributors
+    Copyright ï¿½ 2016 - 2024 Stefan Berg <isbeorn86+NINA@googlemail.com> and the N.I.N.A. contributors
 
     This file is part of N.I.N.A. - Nighttime Imaging 'N' Astronomy.
 
@@ -30,6 +30,9 @@ namespace NINA.Profile.Interfaces {
         bool TimeSync { get; set; }
         bool PrimaryReversed { get; set; }
         bool SecondaryReversed { get; set; }
+        bool OverrideCanPark { get; set; }
+        bool OverrideCanUnpark { get; set; }
+        bool OverrideCanFindHome { get; set; }
 
         TelescopeLocationSyncDirection TelescopeLocationSyncDirection { get; set; }
     }